pyproject.toml
[project]
name = "llm-design-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.82.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipython>=9.2.0",
    "mypy>=1.15.0",
    "pylint>=3.3.7",
    "ruff>=0.11.11",
    "wat-inspector>=0.4.3",
]

[tool.uv.workspace]
members = ["structured_output", "tools", "rag", "pdf_to_text"]
